\docType{data}

\name{cloudms2}

\alias{cloudms2}

\title{Benchmark data set for cloudUtil}

\format{A data frame with 10969 rows and 15 variables}

\description{

The data set provides the measurements of benchmark using three compute 
infrastructers namely uzh.ch Schroedinger, Amazon EC2 (region us-east.), and the 
FGCZ internal compute server. The number crunshing was done by doing tandem mass 
spectra peptide sequence assignments of a large scale proteomics Drosophila 
experiment similar to the data processing published here [E. Brunner, et. al., 
NBT, 2007] having approx. 1800 (LC)-MS/MS runs, 8474960 fragment ion 
spectra(tandem MS or MS2), identidying 498000 redundant and 72281 distinct 
peptides, and 9124 proteins. The data volume is approx. 0.3TB splitted into ~1800 
jobs.

}

\references{

      \url{http://fgcz-transfer.uzh.ch/~cpanse/2011-10-24--pCloud.pdf}

      \url{http://www.nature.com/nbt/journal/v25/n5/abs/nbt1300.html}


}
\keyword{datasets}
